+++ MungDo/Presentation/Screen/Todo/CompletedTask/CompletedTaskView.swift
@@ -174,190 +174,14 @@ struct CompletedTaskView: View {
                         .foregroundColor(Color("ButtonSecondary"))
                 }
             }
-//            ToolbarItem(placement: .principal) {
-//                Text("완료된 할 일")
-//                    .font(.system(size: 20, weight: .semibold))
-//                    .foregroundColor(.black)
-//            }
         }
         .toolbarBackground(Color("BackgroundPrimary"), for: .navigationBar)
         .toolbarBackground(.visible, for: .navigationBar)
     }
 }
 
-struct TaskCardView: View {
-    let task: CompletedTaskView.TaskItem
-    @State private var isPressed = false
 
-    var body: some View {
-        VStack {
-            Spacer()
-            // 이미지 중앙 배치
-            ZStack {
-                Circle()
-                    .fill(Color.pink.opacity(0.1))
-                    .frame(width: 213, height: 213)
-                
-                // 이미지 에셋 사용 (fallback으로 시스템 이미지)
-                if !task.imageName.isEmpty {
-                    Image(task.imageName)
-                        .resizable()
-                        .scaledToFit()
-                        .frame(width: 120, height: 120)
-                        .foregroundColor(.pink)
-                } else {
-                    Image(systemName: "pawprint.fill")
-                        .font(.system(size: 50))
-                        .foregroundColor(.pink)
-                }
-            }
-            // 텍스트는 이미지 바로 아래
-            Text(task.title)
-                .font(.system(size: 27, weight: .semibold))
-                .foregroundColor(.black)
-                .multilineTextAlignment(.center)
-                .lineLimit(3)
-                .padding(.top, 20)
-                .padding(.horizontal, 20)
-            Spacer()
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background(
-            RoundedRectangle(cornerRadius: 24)
-                .fill(Color.white)
-        )
-        .scaleEffect(isPressed ? 0.95 : 1.0)
-        .shadow(color: .black.opacity(0.1), radius: 8, x: 0, y: 4)
-        .onTapGesture {
-            withAnimation(.easeInOut(duration: 0.1)) {
-                isPressed = true
-            }
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                withAnimation(.easeInOut(duration: 0.1)) {
-                    isPressed = false
-                }
-            }
-        }
-    }
-}
 
-struct AllTasksCompletedView: View {
-    let onDismiss: () -> Void
-    @Environment(\.dismiss) private var dismiss
-    
-    var body: some View {
-        ZStack {
-            // 배경
-            Color("BackgroundPrimary")
-                .ignoresSafeArea()
-            
-            VStack(spacing: 40) {
-                Spacer()
-                
-                // 축하 애니메이션과 강아지 일러스트
-                ZStack {
-                    // Confetti 효과
-                    ForEach(0..<20, id: \.self) { index in
-                        ConfettiPiece(index: index)
-                    }
-                    
-                    // 강아지 일러스트 (임시로 시스템 이미지 사용)
-                    VStack(spacing: 20) {
-                        Image(systemName: "dog.fill")
-                            .font(.system(size: 120))
-                            .foregroundColor(.brown)
-                            .scaleEffect(1.2)
-                            .animation(.easeInOut(duration: 1.0).repeatForever(autoreverses: true), value: true)
-                        
-                        // 실제로는 강아지 일러스트 이미지를 사용
-                        // Image("celebration_dog")
-                        //     .resizable()
-                        //     .scaledToFit()
-                        //     .frame(width: 200, height: 200)
-                    }
-                }
-                .frame(height: 300)
-                
-                // 축하 메시지
-                VStack(spacing: 10) {
-                    Text("오늘의 할일을 모두 마쳤어요!")
-                        .font(.system(size: 28, weight: .bold))
-                        .foregroundColor(.black)
-                        .multilineTextAlignment(.center)
-                }
-                
-                Spacer()
-                
-                // 돌아가기 버튼
-                Button(action: {
-                    onDismiss()
-                }) {
-                    Text("돌아가기")
-                        .font(.system(size: 18, weight: .semibold))
-                        .foregroundColor(.white)
-                        .frame(width: 280, height: 60)
-                        .background(
-                            RoundedRectangle(cornerRadius: 30)
-                                .fill(Color("ButtonPrimary"))
-                        )
-                        .shadow(color: .pink.opacity(0.3), radius: 10, x: 0, y: 5)
-                }
-                .padding(.bottom, 50)
-            }
-            
-            // 오른쪽 상단 X 버튼
-            VStack {
-                HStack {
-                    Spacer()
-                    Button(action: {
-                        onDismiss()
-                    }) {
-                        Image(systemName: "xmark")
-                            .font(.system(size: 24, weight: .medium))
-                            .foregroundColor(.gray)
-                            .frame(width: 44, height: 44)
-                            .background(
-                                Circle()
-                                    .fill(Color.white.opacity(0.8))
-                            )
-                    }
-                    .padding(.top, 60)
-                    .padding(.trailing, 30)
-                }
-                Spacer()
-            }
-        }
-    }
-}
-
-// Confetti 효과를 위한 컴포넌트 (더 간단한 버전)
-struct ConfettiPiece: View {
-    let index: Int
-    @State private var isAnimating = false
-    
-    var body: some View {
-        let colors: [Color] = [.red, .blue, .green, .yellow, .orange, .purple, .pink]
-        let color = colors[index % colors.count]
-        
-        Rectangle()
-            .fill(color)
-            .frame(width: 8, height: 8)
-            .rotationEffect(.degrees(isAnimating ? 360 : 0))
-            .offset(
-                x: isAnimating ? CGFloat.random(in: -200...200) : 0,
-                y: isAnimating ? CGFloat.random(in: -300...300) : 0
-            )
-            .opacity(isAnimating ? 0 : 1)
-            .onAppear {
-                withAnimation(
-                    .easeOut(duration: Double.random(in: 2...4))
-                    .delay(Double.random(in: 0...2))
-                ) {
-                    isAnimating = true
-                }
-            }
-    }
-}
 
 #Preview {
     NavigationStack {
